Student can now Select New Courses Available under department in which student is registered.
Students can now login and go to select new courses activity where a list view is present in which courses which are available in the department in which student is registered are shown.
All the course details such as course code, name, department , professor are shown , and there is check box to select and unselect a particular course.
When student select check box its details are updated in firabase realtime database.